pyproject.toml
[tool.poetry]
name = "ndbioimage"
version = "2023.11.0"
description = "Bio image reading, metadata and some affine registration."
authors = ["W. Pomp <w.pomp@nki.nl>"]
license = "GPLv3"
readme = "README.md"
keywords = ["bioformats", "imread", "numpy", "metadata"]
include = ["transform.txt"]
repository = "https://github.com/wimpomp/ndbioimage"
exclude = [ "ndbioimage/jars" ]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "*"
pandas = "*"
tifffile = "*"
czifile = "*"
tiffwrite = "*"
ome-types = "^0.4.0"
pint = "*"
tqdm = "*"
lxml = "*"
pyyaml = "*"
parfor = ">=2023.10.1"
JPype1 = "*"
SimpleITK-SimpleElastix = "*"
scikit-image = "*"
pytest = { version = "*", optional = true }

[tool.poetry.extras]
test = ["pytest-xdist"]

[tool.poetry.scripts]
ndbioimage = "ndbioimage:main"

[tool.pytest.ini_options]
filterwarnings = ["ignore:::(colorcet)"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
